HEJAHEJA

+++ utvecklingsprojekt/articles.php
@@ -30,9 +30,9 @@ END;
 		if(isset($_SESSION['userId']))
 		{
 			    $content .= <<<END
-					<a href="delete.php?artikelnr={$row->artikelnr}" onclick="return
-				confirm('Är du säker?')">Ta bort artikel</a><br>
-				<a href="edit_article.php?artikelnr={$row->artikelnr}">Redigera artikel</a></br>
+			    <a href="edit_article.php?artikelnr={$row->artikelnr}">Redigera artikel</a><br>
+			    <a href="delete.php?artikelnr={$row->artikelnr}="deletelink" onclick="return confirm('Are you sure?')">Radera artikel</a></br>
+				
 END;
 		}
